Aded Methods for Notice and Academic CRUDs

+++ ProfessionalPracticesSystem/BusinessLogic/ManagePractitioner.cs
@@ -46,7 +46,10 @@ namespace BusinessLogic
         {
             bool isDeleted = false;
 
-            isDeleted = practitionerDao.DeletePractitioner(deletedPractitionerID);
+            if (deletedPractitionerID != INVALID_ID)
+            {
+                isDeleted = practitionerDao.DeletePractitioner(deletedPractitionerID);
+            } 
 
             return isDeleted;
         }
